+++ src/egon/data/airflow/dags/pipeline.py
@@ -10,12 +10,16 @@ from egon.data.datasets.data_bundle import DataBundle
 from egon.data.datasets.era5 import WeatherData
 from egon.data.datasets.heat_etrago import HeatEtrago
 from egon.data.datasets.heat_supply import HeatSupply
+from egon.data.datasets.industry import IndustrialDemandCurves
+from egon.data.datasets.industrial_sites import MergeIndustrialSites
 from egon.data.datasets.renewable_feedin import RenewableFeedin
 from egon.data.datasets.osm import OpenStreetMap
 from egon.data.datasets.mastr import mastr_data_setup
 from egon.data.datasets.re_potential_areas import re_potential_area_setup
 from egon.data.datasets.mv_grid_districts import mv_grid_districts_setup
 from egon.data.datasets.power_plants import PowerPlants
+from egon.data.datasets.scenario_capacities import ScenarioCapacities
+from egon.data.datasets.scenario_parameters import ScenarioParameters
 from egon.data.datasets.vg250 import Vg250
 from egon.data.processing.zensus_vg250 import (
     zensus_population_inside_germany as zensus_vg250,
@@ -25,10 +29,6 @@ import egon.data.importing.demandregio as import_dr
 import egon.data.importing.demandregio.install_disaggregator as install_dr
 import egon.data.importing.etrago as etrago
 import egon.data.importing.heat_demand_data as import_hd
-import egon.data.importing.industrial_sites as industrial_sites
-
-import egon.data.importing.nep_input_data as nep_input
-import egon.data.importing.scenarios as import_scenarios
 import egon.data.importing.zensus as import_zs
 import egon.data.importing.gas_grid as gas_grid
 
@@ -36,11 +36,10 @@ import egon.data.processing.boundaries_grid_districts as boundaries_grid_distric
 import egon.data.processing.demandregio as process_dr
 import egon.data.processing.district_heating_areas as district_heating_areas
 import egon.data.processing.osmtgmod as osmtgmod
+import egon.data.processing.power2gas as power2gas
 import egon.data.processing.substation as substation
-import egon.data.processing.zensus_vg250.zensus_population_inside_germany as zensus_vg250
 import egon.data.processing.gas_areas as gas_areas
 import egon.data.importing.scenarios as import_scenarios
-import egon.data.importing.industrial_sites as industrial_sites
 import egon.data.processing.loadarea as loadarea
 import egon.data.processing.calculate_dlr as dlr
 
@@ -85,6 +84,10 @@ with airflow.DAG(
     vg250 = Vg250(dependencies=[setup])
     vg250.insert_into(pipeline)
     vg250_clean_and_prepare = tasks["vg250.cleaning-and-preperation"]
+
+    # Scenario table
+    scenario_parameters = ScenarioParameters(dependencies=[setup])
+    scenario_input_import = tasks["scenario_parameters.insert-scenarios"]
 
     # Zensus import
     zensus_download_population = PythonOperator(
@@ -147,26 +150,13 @@ with airflow.DAG(
     ] >> map_zensus_vg250 >> zensus_inside_ger >> zensus_inside_ger_metadata
     zensus_inside_ger >> vg250_population >> vg250_population_metadata
 
-    # Scenario table
-    scenario_input_tables = PythonOperator(
-        task_id="create-scenario-parameters-table",
-        python_callable=import_scenarios.create_table
-    )
-
-    scenario_input_import = PythonOperator(
-        task_id="import-scenario-parameters",
-        python_callable=import_scenarios.insert_scenarios
-    )
-    setup >> scenario_input_tables >> scenario_input_import
-
     # DemandRegio data import
     demandregio_tables = PythonOperator(
         task_id="demandregio-tables",
         python_callable=import_dr.create_tables,
     )
 
-    scenario_input_tables >> demandregio_tables
-
+    scenario_input_import >> demandregio_tables
 
     demandregio_installation = PythonOperator(
         task_id="demandregio-installation",
@@ -251,20 +241,11 @@ with airflow.DAG(
     map_zensus_vg250 >> elec_household_demands_zensus
 
     # NEP data import
-    create_tables = PythonOperator(
-        task_id="create-scenario-tables",
-        python_callable=nep_input.create_scenario_input_tables,
-    )
+    scenario_capacities = ScenarioCapacities(
+        dependencies=[setup, vg250, data_bundle])
+    nep_insert_data = tasks["scenario_capacities.insert-data-nep"]
 
-    nep_insert_data = PythonOperator(
-        task_id="insert-nep-data",
-        python_callable=nep_input.insert_data_nep,
-    )
-
-    setup >> create_tables >> nep_insert_data
-    vg250_clean_and_prepare >> nep_insert_data
     population_import >> nep_insert_data
-    download_data_bundle >> nep_insert_data
 
     # setting etrago input tables
     etrago_input_data = PythonOperator(
@@ -366,26 +347,6 @@ with airflow.DAG(
     zensus_inside_ger_metadata >> heat_demand_import
     scenario_input_import >> heat_demand_import
 
-
-    # Import and merge data on industrial sites from different sources
-
-    industrial_sites_import = PythonOperator(
-        task_id="download-import-industrial-sites",
-        python_callable=industrial_sites.download_import_industrial_sites
-    )
-
-    industrial_sites_merge = PythonOperator(
-        task_id="merge-industrial-sites",
-        python_callable=industrial_sites.merge_inputs
-    )
-
-    industrial_sites_nuts = PythonOperator(
-        task_id="map-industrial-sites-nuts3",
-        python_callable=industrial_sites.map_nuts3
-    )
-    vg250_clean_and_prepare >> industrial_sites_import
-    industrial_sites_import >> industrial_sites_merge >> industrial_sites_nuts
-
     # Distribute electrical CTS demands to zensus grid
 
     elec_cts_demands_zensus = PythonOperator(
@@ -398,7 +359,6 @@ with airflow.DAG(
     demandregio_demand_cts_ind >> elec_cts_demands_zensus
     map_zensus_vg250 >> elec_cts_demands_zensus
 
-
     # Gas grid import
     gas_grid_insert_data = PythonOperator(
         task_id="insert-gas-grid",
@@ -407,6 +367,16 @@ with airflow.DAG(
 
     etrago_input_data >> gas_grid_insert_data
     download_data_bundle >> gas_grid_insert_data
+    osmtgmod_pypsa >> gas_grid_insert_data
+
+     # Power-to-gas installations creation
+    insert_power2gas_installations = PythonOperator(
+        task_id="insert-power-to-gas-installations",
+        python_callable=power2gas.insert_power2gas,
+    )
+
+    gas_grid_insert_data >> insert_power2gas_installations
+    osmtgmod_pypsa >> insert_power2gas_installations
 
     # Create gas voronoi
     create_gas_polygons = PythonOperator(
@@ -511,6 +481,7 @@ with airflow.DAG(
     hvmv_substation_extraction >> generate_pv_ground_mounted
     nep_insert_data >> power_plant_import
     map_boundaries_grid_districts >> solar_rooftop_etrago
+    feedin_pv >> solar_rooftop_etrago
     elec_cts_demands_zensus >> solar_rooftop_etrago
     elec_household_demands_zensus >> solar_rooftop_etrago
     nep_insert_data >> solar_rooftop_etrago
@@ -542,3 +513,13 @@ with airflow.DAG(
     heat_etrago_supply = tasks["heat_etrago.supply"]
 
     etrago_input_data >> heat_etrago_buses
+
+    # Industry
+
+    industrial_sites = MergeIndustrialSites(
+        dependencies=[setup, vg250_clean_and_prepare]
+    )
+
+    demand_curves_industry = IndustrialDemandCurves(
+        dependencies=[define_mv_grid_districts, industrial_sites, demandregio_demand_cts_ind, osm, landuse_extraction]
+    )